Added functionality and changes to navigation
When the user grabs an item, information about it will be displayed until the user lets the item go. The user can now decide if they want to go to the museum via the title screen or after visiting a room. The 'Images' folder was put into the 'Resources' folder